Moving some serialized data to their own file
The goal was to fix test depending on some nasty global state but that's not enough.

I will need a separate PR for that
