+++ src/ts/missions/generateSightSeeingMissions.ts
@@ -20,7 +20,8 @@ import { newSightSeeingMission } from "./sightSeeingMission";
 import { uniformRandBool } from "extended-random";
 import { SystemObjectType } from "../utils/coordinates/universeCoordinates";
 import { Player } from "../player/player";
-import { Mission, MissionType } from "./mission";
+import { Mission } from "./mission";
+import { MissionType } from "./missionSerialized";
 import { StarSystemModel, StarSystemModelUtils } from "../starSystem/starSystemModel";
 import { getRngFromSeed } from "../utils/getRngFromSeed";
 import { OrbitalObjectType } from "../architecture/orbitalObjectType";
