+++ src/ts/missions/mission.ts
@@ -17,35 +17,11 @@
 
 import { MissionContext } from "./missionContext";
 import i18n from "../i18n";
-import {
-    StarSystemCoordinates,
-    UniverseObjectId,
-    UniverseObjectIdSchema
-} from "../utils/coordinates/universeCoordinates";
-import { deserializeMissionNode, MissionNode, MissionNodeSerializedSchema } from "./nodes/deserializeNode";
+import { StarSystemCoordinates, UniverseObjectId } from "../utils/coordinates/universeCoordinates";
+import { deserializeMissionNode } from "./nodes/deserializeNode";
+import { MissionNode } from "./nodes/missionNode";
 import { StarSystemDatabase } from "../starSystem/starSystemDatabase";
-import { z } from "zod";
-
-/**
- * Registered mission types. Those are used to display localized strings in the UI
- */
-export enum MissionType {
-    SIGHT_SEEING_FLY_BY,
-    SIGHT_SEEING_TERMINATOR_LANDING,
-    SIGHT_SEEING_ASTEROID_FIELD
-}
-
-export const MissionSerializedSchema = z.object({
-    missionGiver: UniverseObjectIdSchema,
-    tree: MissionNodeSerializedSchema,
-    reward: z.number().default(0),
-    type: z.nativeEnum(MissionType).default(MissionType.SIGHT_SEEING_FLY_BY)
-});
-
-/**
- * Serialized mission object as stored in save files
- */
-export type MissionSerialized = z.infer<typeof MissionSerializedSchema>;
+import { MissionType, MissionSerialized } from "./missionSerialized";
 
 /**
  * General mission abstraction. The mission can have any arbitrary task tree and reward.
