+++ src/ts/missions/nodes/actions/sightseeing/missionAsteroidFieldNode.ts
@@ -32,7 +32,7 @@ import { parseDistance } from "../../../../utils/strings/parseToStrings";
 import { Settings } from "../../../../settings";
 import { getGoToSystemInstructions } from "../../../common";
 import { StarSystemDatabase } from "../../../../starSystem/starSystemDatabase";
-import type { MissionNode } from "../../deserializeNode";
+import type { MissionNode } from "../../missionNode";
 import { z } from "zod";
 
 enum AsteroidFieldMissionState {
