+++ src/ts/missions/nodes/logic/missionOrNode.ts
@@ -20,7 +20,8 @@ import { MissionContext } from "../../missionContext";
 import i18n from "../../../i18n";
 import { StarSystemCoordinates } from "../../../utils/coordinates/universeCoordinates";
 import { StarSystemDatabase } from "../../../starSystem/starSystemDatabase";
-import type { MissionNode, MissionNodeSerialized } from "../deserializeNode";
+import type { MissionNode } from "../missionNode";
+import type { MissionNodeSerialized } from "../missionNodeSerialized";
 
 export type MissionOrNodeSerialized = MissionNodeSerializedBase<MissionNodeType.OR> & {
     children: MissionNodeSerialized[];
