+++ src/ts/missions/nodes/missionNodeSerialized.ts
@@ -1,0 +1,75 @@
+//  This file is part of Cosmos Journeyer
+//
+//  Copyright (C) 2024 Barthélemy Paléologue <barth.paleologue@cosmosjourneyer.com>
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import { z } from "zod";
+import {
+    MissionAsteroidFieldNodeSerialized,
+    MissionAsteroidFieldNodeSerializedSchema
+} from "./actions/sightseeing/missionAsteroidFieldNode";
+import { MissionFlyByNodeSerialized, MissionFlyByNodeSerializedSchema } from "./actions/sightseeing/missionFlyByNode";
+import {
+    MissionTerminatorLandingNodeSerialized,
+    MissionTerminatorLandingNodeSerializedSchema
+} from "./actions/sightseeing/missionTerminatorLandingNode";
+import { MissionAndNodeSerialized } from "./logic/missionAndNode";
+import { MissionOrNodeSerialized } from "./logic/missionOrNode";
+import { MissionSequenceNodeSerialized } from "./logic/missionSequenceNode";
+import { MissionXorNodeSerialized } from "./logic/missionXorNode";
+import { MissionNodeType } from "./missionNode";
+
+export type MissionNodeSerialized =
+    | MissionAndNodeSerialized
+    | MissionOrNodeSerialized
+    | MissionXorNodeSerialized
+    | MissionSequenceNodeSerialized
+    | MissionAsteroidFieldNodeSerialized
+    | MissionFlyByNodeSerialized
+    | MissionTerminatorLandingNodeSerialized;
+
+export const MissionNodeSerializedSchema: z.ZodType<MissionNodeSerialized, z.ZodTypeDef, MissionNodeSerialized> =
+    z.lazy(() =>
+        z.discriminatedUnion("type", [
+            MissionFlyByNodeSerializedSchema,
+            MissionTerminatorLandingNodeSerializedSchema,
+            MissionAsteroidFieldNodeSerializedSchema,
+            MissionAndNodeSerializedSchema,
+            MissionOrNodeSerializedSchema,
+            MissionXorNodeSerializedSchema,
+            MissionSequenceNodeSerializedSchema
+        ])
+    );
+
+export const MissionOrNodeSerializedSchema = z.object({
+    type: z.literal(MissionNodeType.OR),
+    children: z.array(MissionNodeSerializedSchema)
+});
+
+export const MissionAndNodeSerializedSchema = z.object({
+    type: z.literal(MissionNodeType.AND),
+    children: z.array(MissionNodeSerializedSchema)
+});
+
+export const MissionXorNodeSerializedSchema = z.object({
+    type: z.literal(MissionNodeType.XOR),
+    children: z.array(MissionNodeSerializedSchema)
+});
+
+export const MissionSequenceNodeSerializedSchema = z.object({
+    type: z.literal(MissionNodeType.SEQUENCE),
+    activeChildIndex: z.number(),
+    children: z.array(MissionNodeSerializedSchema)
+});
